+++ simulator/short-circuit/src/main/java/com/powsybl/incubator/simulator/shortcircuit/AbstractShortCircuitEngine.java
@@ -13,9 +13,7 @@ import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.ThreeWindingsTransformer;
 import com.powsybl.incubator.simulator.util.AdmittanceEquationSystem;
 import com.powsybl.incubator.simulator.util.CalculationLocation;
-import com.powsybl.incubator.simulator.util.ShortCircuitFault;
 import com.powsybl.incubator.simulator.util.extensions.ShortCircuitExtensions;
-import com.powsybl.incubator.simulator.util.ShortCircuitResult;
 import com.powsybl.openloadflow.OpenLoadFlowParameters;
 import com.powsybl.openloadflow.ac.outerloop.AcLoadFlowParameters;
 import com.powsybl.openloadflow.graph.EvenShiloachGraphDecrementalConnectivityFactory;
