+++ simulator/short-circuit/src/main/java/com/powsybl/incubator/simulator/shortcircuit/OpenShortCircuitProvider.java
@@ -9,9 +9,6 @@ package com.powsybl.incubator.simulator.shortcircuit;
 import com.google.auto.service.AutoService;
 import com.google.common.base.Stopwatch;
 import com.powsybl.incubator.simulator.util.extensions.AdditionalDataInfo;
-import com.powsybl.incubator.simulator.util.ShortCircuitFault;
-import com.powsybl.incubator.simulator.util.ShortCircuitNormCourcirc;
-import com.powsybl.incubator.simulator.util.ShortCircuitResult;
 import com.powsybl.computation.ComputationManager;
 import com.powsybl.iidm.network.Bus;
 import com.powsybl.iidm.network.Network;
