+++ simulator/short-circuit/src/main/java/com/powsybl/incubator/simulator/shortcircuit/ShortCircuitEngineParameters.java
@@ -6,9 +6,7 @@
  */
 package com.powsybl.incubator.simulator.shortcircuit;
 
-import com.powsybl.incubator.simulator.util.ShortCircuitFault;
 import com.powsybl.incubator.simulator.util.extensions.AdditionalDataInfo;
-import com.powsybl.incubator.simulator.util.ShortCircuitNorm;
 import com.powsybl.loadflow.LoadFlowParameters;
 import com.powsybl.math.matrix.MatrixFactory;
