+++ simulator/short-circuit/src/main/java/com/powsybl/incubator/simulator/shortcircuit/ShortCircuitFault.java
@@ -4,7 +4,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.incubator.simulator.util;
+package com.powsybl.incubator.simulator.shortcircuit;
+
+import com.powsybl.incubator.simulator.util.CalculationLocation;
 
 /**
  * @author Jean-Baptiste Heyberger <jbheyberger at gmail.com>
